Update useragent, store it in defaults
Useragent updated to Safari 10.1 on macOS 10.12.4
Also, it's now stored in userdefaults, with key userAgent, so it would
be possible to update it without updating the app.

Sidenote: Due to way UserSettings are implemented now, it won't be easy
to update the useragent in future. There has to be some migration
process